Reorg and cleanup
