--- lib/src/optimism/epoch.rs
@@ -1,34 +1,0 @@
-// Copyright 2023 RISC Zero, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use serde::{Deserialize, Serialize};
-use zeth_primitives::{
-    block::Header,
-    receipt::Receipt,
-    transactions::{ethereum::EthereumTxEssence, Transaction, TxEssence},
-};
-
-/// Input for the L2 derivation.
-pub type Input = Vec<BlockInput<EthereumTxEssence>>;
-
-/// Input for extracting deposits.
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct BlockInput<E: TxEssence> {
-    /// Header of the block.
-    pub block_header: Header,
-    /// Transactions of the block.
-    pub transactions: Vec<Transaction<E>>,
-    /// Transaction receipts of the block or `None` if not required.
-    pub receipts: Option<Vec<Receipt>>,
-}
